feat: PHP JsonSerializable Preset
In PHP objects (and also arrays and scalars) can be serialized to JSON
using the function `json_encode()`. In order to maintain the correct
property names and values for Enums, normalization of values needs to be
implemented manually using the `JsonSerializable` interface.

The serialization to JSON can be enabled by using the optional preset
`PHP_JSON_SERIALIZABLE_PRESET`.

This way of serializing PHP objects does not require any user land
package and works out of the box with each PHP version. It is also
compatible with different serialization frameworks, like Symfony.