pydantic generator: support dash in property names
This enables the generated models for Pydantic to work for properties
with dashes in the name.

The generated code now puts descriptions in the "description" argument
of the Pydantic Field decorator, and puts the original, unconstrained,
property name in the "alias" field. This is correct usage as per the
[Pydantic documentation](https://docs.pydantic.dev/latest/concepts/fields/#field-aliases)

Also, dashes are allowed through in property key constrainer, so that a
property name like "content-type" is converted into "contentType" instead
of "contentMinusType".